… services/frontend/package.json
… services/frontend/webpack.config.js