https://rsbuild.dev/

`Compiled in 1.09 s` vs `webpack 5.93.0 compiled with 2 warnings in 11016 ms`


```
> rsbuild build --env-mode=production

  Rsbuild v1.0.1-beta.9
ready   Compiled in 1.09 s (web)
info    Production file sizes for web:

  File                                     Size        Gzipped    
  dist/favicon.ico                         4.3 kB      1.3 kB
  dist/index.html                          10.8 kB     2.7 kB
  dist/static/js/index.e671be8d.js         16.4 kB     6.2 kB
  dist/static/js/async/pako.bc65a9a4.js    46.7 kB     14.5 kB
  dist/static/css/582.f657d9c9.css         64.0 kB     4.1 kB
  dist/static/js/582.aa191e11.js           486.1 kB    152.5 kB
  dist/static/css/index.ebbc67c9.css       841.9 kB    72.5 kB

  Total size:  1470.2 kB
  Gzipped size:  253.8 kB
```

vs

```
> webpack --mode=production

assets by info 1.27 MiB [immutable]
  assets by path *.js 535 KiB
    asset vendors.bee6b306a0cfbb0be63e.js 473 KiB [emitted] [immutable] [minimized] [big] (name: vendors) (id hint: vendor) 1 related asset
    asset pako.ceea69992014b2d8ef7a.js 45.6 KiB [emitted] [immutable] [minimized] (name: pako) (id hint: pako, vendors) 1 related asset
    asset main.fc8ce41033dfb6af7a48.js 13.5 KiB [emitted] [immutable] [minimized] (name: main) 1 related asset
    asset runtime.445917f1f30b62621ff2.js 3.11 KiB [emitted] [immutable] [minimized] (name: runtime) 1 related asset
  assets by path *.css 769 KiB
    asset main.41bbad6148f33de6c2d8.css 704 KiB [emitted] [immutable] [minimized] [big] (name: main) 1 related asset
    asset vendors.00cf026c4eaf2e521ec2.css 64.8 KiB [emitted] [immutable] [minimized] (name: vendors) (id hint: vendor) 1 related asset
asset index.html 6.4 KiB [emitted]
asset favicon.ico 4.19 KiB [compared for emit] [from: src/static/favicon.ico] [copied]
Entrypoint main [big] 1.23 MiB (3.97 MiB) = runtime.445917f1f30b62621ff2.js 3.11 KiB vendors.00cf026c4eaf2e521ec2.css 64.8 KiB vendors.bee6b306a0cfbb0be63e.js 473 KiB main.41bbad6148f33de6c2d8.css 704 KiB main.fc8ce41033dfb6af7a48.js 13.5 KiB 5 auxiliary assets
orphan modules 1.53 MiB (javascript) 937 bytes (runtime) [orphan] 213 modules
runtime modules 8.01 KiB 12 modules
cacheable modules 2.97 MiB (javascript) 802 KiB (css/mini-extract)
  modules by path ./node_modules/rxjs/dist/esm5/internal/ 78.8 KiB 50 modules
  modules by path ./node_modules/@codemirror/ 936 KiB 9 modules
  modules by path ./node_modules/@fortawesome/ 1.59 MiB
    ./node_modules/@fortawesome/fontawesome-svg-core/index.mjs 90.6 KiB [built] [code generated]
    + 2 modules
  modules by path ./src/ 28.7 KiB (javascript) 718 KiB (css/mini-extract)
    ./src/ts/index.ts + 8 modules 28.7 KiB [built] [code generated]
    css ./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js??ruleSet[1].rules[1].use[2]!./src/resources/css/style.scss 718 KiB [built] [code generated]
  modules by path ./node_modules/@lezer/ 107 KiB
    ./node_modules/@lezer/common/dist/index.js 79.1 KiB [built] [code generated]
    ./node_modules/@lezer/highlight/dist/index.js 28.2 KiB [built] [code generated]
  + 6 modules

WARNING in asset size limit: The following asset(s) exceed the recommended size limit (244 KiB).
This can impact web performance.
Assets: 
  main.41bbad6148f33de6c2d8.css (704 KiB)
  vendors.bee6b306a0cfbb0be63e.js (473 KiB)

WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
Entrypoints:
  main (1.23 MiB)
      runtime.445917f1f30b62621ff2.js
      vendors.00cf026c4eaf2e521ec2.css
      vendors.bee6b306a0cfbb0be63e.js
      main.41bbad6148f33de6c2d8.css
      main.fc8ce41033dfb6af7a48.js


webpack 5.93.0 compiled with 2 warnings in 11016 ms
```